Fix `JSONObject["text"] not a string` error
I ran into a bug where the program would crash on certain messages that don't have text associated with them. The JSON would contain `{"text": null}`, but `message.has("text")` would still return `true`, as the value would be a `JSONObject.NULL` object. Therefore I added another check for that. I didn't just add this check for `text`, but for all keys in this file, to harden it a bit better than before.